S3 Client support for R66
Use MinIO client for integration

Only from Java 8 version support: Special WaarpR66-S3 Jars for JRE 8 and JRE 11 is supported.

Make integration within R66 within the new module WaarpS3 as new Tasks using EXTENDED command.

- Task : S3GET get the remote file with options
- Task : S3GETDELETE get the remote file with options and with delete remote file at the end
- Task : S3PUT create remote file with options
- Task : S3PUTR66DELETE create remote file with options and with delete local file at the end
- Task : S3DELETE delete remote file at the end

The Rule shall be using `SENDMD5THROUGHMODE` or `SENDTHROUGHMODE` mode
(respectively 7 or 5) or `RECVMD5THROUGHMODE` or `RECVTHROUGHMODE` mode
(respectively 8 or 6) as pre-tasks since the file does not exist at
startup when file is to be retrived first through S3GET or S3GETDELETE.

Junit tests added for WaarpS3Client and for all S3 Tasks.

Also fix changing default.buffer to default.ioBuffer when possible and wrappedBuffers,
and fix R66Proxy.

+++ WaarpDigest/src/main/java/org/waarp/common/digest/MD5.java
@@ -780,7 +780,7 @@ class MD5 {
    **/
   static byte[] getHash(final InputStream stream) throws IOException {
     try {
-      final byte[] buf = new byte[(int) ZERO_COPY_CHUNK_SIZE];
+      final byte[] buf = new byte[ZERO_COPY_CHUNK_SIZE];
       final MD5 md5 = new MD5();
       int read;
       while ((read = stream.read(buf)) >= 0) {
